Hice el ejercicio de practica para el corto. Cree una carpeta para ejercicos de clase y movi el ejercicio de NotaLetra de la carpeta for a la nueva carpeta

+++ Switch/Calculadora.cpp
@@ -4,7 +4,7 @@ using namespace std;
 int main()
 {
 
-    float num1, num2, resultado;
+    double num1, num2, resultado;
     int opcion;
 
     cout << "ingrese el primer numero" << endl;
